# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/categories": {
      "get": {
        "description": "Retrieves a paginated list of categories with search and sorting.",
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "enum": [
                "name",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query for category name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully"
          },
          "400": {
            "description": "Invalid query parameters"
          }
        },
        "summary": "List all categories",
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "description": "Creates a new category. Admin only.",
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "Fiction"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "example": "Fictional literature including novels and short stories"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully"
          },
          "400": {
            "description": "Invalid input data or category already exists"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/categories/{id}": {
      "patch": {
        "description": "Updates an existing category. Admin only.",
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "description": "Deletes a category if not referenced by any books. Admin only.",
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category is referenced by books and cannot be deleted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Management System API",
    "description": "RESTful API for a comprehensive Library Management System. This API provides endpoints for managing books, authors, categories, book copies, members, loans, and system settings.",
    "version": "1.0.0",
    "contact": {
      "name": "Library Management Team",
      "url": "https://github.com/library-management",
      "email": "support@library-management.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.library-management.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}